+++ part2/coursecontents/src/components/course.jsx
@@ -1,0 +1,8 @@
+import React from 'react';
+
+const course = () => {
+    return "Course Content here";
+}
+
+
+export default course;
